sc_download/requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
get==2019.4.13
html5lib==1.0.1
idna==2.8
lxml==4.4.1
post==2019.4.13
public==2019.4.13
python-crontab==2.4.0
python-dateutil==2.8.1
query-string==2019.4.13
requests==2.22.0
six==1.12.0
soupsieve==1.9.4
urllib3==1.25.6
webencodings==0.5.1
xlrd==1.2.0
